backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
packaging==26.2
peewee==4.0.5
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
pwdlib==0.3.0
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pyrefly==0.63.1
pytest==9.0.3
python-dotenv==1.2.2
SQLAlchemy==2.0.49
starlette==1.0.0
types-peewee==4.0.1.20260426
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
